fix: Uso de custom errors, corrección de sesión en CRUD tareas

+++ src/modules/task/interface/task.service.ts
@@ -1,60 +1,57 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from '../../../prisma.service';
 import { CreateTaskDto } from '../dto/create-task.dto';
 import { UpdateTaskDto } from '../dto/update.task.dto';
+import { Task } from '../entities/task.entity';
 
 @Injectable()
 export class TaskService {
-  constructor(private prisma: PrismaService) {}
 
-  async getTasks() {
-    return this.prisma.task.findMany();
+  constructor(@Inject('PG_CONNECTION') private db: any, private prisma: PrismaService) { }
+
+  private tasks: any[] = [];
+
+  async getTasks(user_id: number): Promise<Task[]> {
+    const tasks = await this.prisma.task.findMany({
+      where: {
+        user_id
+      }
+    });
+    return tasks;
   }
 
-  async getTaskById(id: number) {
-    const task = await this.prisma.task.findUnique({ where: { id } });
-    if (!task) throw new NotFoundException(`Tarea con ID ${id} no encontrada`);
+  async getTaskById(id: number, user_id: number): Promise<Task> {
+    const task = await this.prisma.task.findUnique({
+      where: { id, user_id },
+    });
+
+    if (task == undefined) {
+      throw new NotFoundException(`Tarea con ID ${id} no encontrada`);
+    }
+
     return task;
   }
 
-  async insertTask(dto: CreateTaskDto) {
-    return this.prisma.task.create({
+  async insertTask(task: CreateTaskDto, user_id: number): Promise<Task> {
+    const newTask = await this.prisma.task.create({
       data: {
-        name: dto.name,
-        description: dto.description,
-        priority: dto.priority,
-        userId: dto.user_id,
-      },
+        ...task,
+        user_id
+      }
     });
+
+    return newTask;
   }
 
-  async updateTask(id: number, dto: UpdateTaskDto) {
-  await this.getTaskById(id);
-  return this.prisma.task.update({
-    where: { id },
-    data: {
-      name: dto.name,
-      description: dto.description,
-      priority: dto.priority,
-      userId: dto.user_id,
-    },
-  });
-}
+  async updateTask(id: number, user_id: number, taskUpdate: UpdateTaskDto): Promise<Task> {
+    const task = await this.prisma.task.update({ where: { id, user_id }, data: taskUpdate });
 
-  async deleteTask(id: number): Promise<boolean> {
-    try {
-      await this.prisma.task.delete({ where: { id } });
-      return true;
-    } catch {
-      return false;
-    }
+    return task;
   }
 
-  async getTasksByUser(userId: number) {
-  await this.prisma.user.findUniqueOrThrow({ where: { id: userId } }).catch(() => {
-    throw new NotFoundException(`Usuario con ID ${userId} no encontrado`);
-  });
+  async deleteTask(id: number, user_id:number): Promise<boolean> {
+    const deletedTask = await this.prisma.task.delete({ where: { id, user_id } });
 
-  return this.prisma.task.findMany({ where: { userId } });
-}
+    return deletedTask ? true : false;
+  }
 }
